// src/types/beats/sui.ts
export interface SuiCheckpointTransactionsTransaction {
  data: SuiCheckpointTransactionsTransactionData;
  txSignatures: string[];
}
// ...
export interface SuiCheckpointTransactionsEffects {
  messageVersion: string;
  status: SuiCheckpointTransactionsEffectsStatus;
  executedEpoch: string;
  gasUsed: SuiCheckpointTransactionsEffectsGasused;
  modifiedAtVersions: SuiCheckpointTransactionsEffectsModifiedatversions[];
  sharedObjects?: SuiCheckpointTransactionsEffectsSharedobjects[];
  transactionDigest: string;
  mutated: SuiCheckpointTransactionsEffectsMutated[];
  gasObject: SuiCheckpointTransactionsEffectsGasobject;
  dependencies: string[];
  eventsDigest?: string;
  created?: SuiCheckpointTransactionsEffectsCreated[];
  deleted?: SuiCheckpointTransactionsEffectsDeleted[];
  abortError?: SuiCheckpointTransactionsEffectsAborterror;
}
// ...
export interface SuiCheckpointTransactionsEvents {
  id: SuiCheckpointTransactionsEventsId;
  packageId: string;
  transactionModule: string;
  sender: string;
  type: string;
  parsedJson: SuiCheckpointTransactionsEventsParsedjson;
  bcsEncoding: string;
  bcs: string;
}

export interface SuiCheckpointTransactionsObjectchangesOwnerShared {
  initial_shared_version: number;
}

export interface SuiCheckpointTransactionsObjectchangesOwner {
  Shared?: SuiCheckpointTransactionsObjectchangesOwnerShared;
  AddressOwner?: string;
  ObjectOwner?: string;
}

export interface SuiCheckpointTransactionsObjectchanges {
  type: string;
  sender: string;
  owner: SuiCheckpointTransactionsObjectchangesOwner;
  objectType: string;
  objectId: string;
  version: string;
  previousVersion: string;
  digest: string;
}

export interface SuiCheckpointTransactionsBalancechangesOwner {
  AddressOwner: string;
}

export interface SuiCheckpointTransactionsBalancechanges {
  owner: SuiCheckpointTransactionsBalancechangesOwner;
  coinType: string;
  amount: string;
}

export interface SuiCheckpointTransactions {
  digest: string;
  transaction: SuiCheckpointTransactionsTransaction;
  rawTransaction: string;
  effects: SuiCheckpointTransactionsEffects;
  events: SuiCheckpointTransactionsEvents[];
  objectChanges: SuiCheckpointTransactionsObjectchanges[];
  balanceChanges: SuiCheckpointTransactionsBalancechanges[];
  timestampMs: string;
  checkpoint: string;
  rawEffects: number[];
}

export interface SuiCheckpoint {
  sequence: string;
  digest: string;
  timestamp: string;
  transactions: SuiCheckpointTransactions[];
  _network: string;
  _vm: string;
}